requirements.txt
streamlit==1.35.0
opencv-python-headless==4.9.0.80
numpy==1.24.4
Pillow==10.3.0
setuptools>=69.0
wheel
build